Move kundu_tedpca into selection and rename select_comps to tedica.
