+++ tedana/selection/__init__.py
@@ -1,8 +1,7 @@
 # emacs: -*- mode: python-mode; py-indent-offset: 4; tab-width: 4; indent-tabs-mode: nil -*-
 # ex: set sts=4 ts=4 sw=4 et:
 
-from .select_comps import (
-    kundu_selection_v2, manual_selection,
-)
+from .tedpca import kundu_tedpca
+from .tedica import kundu_selection_v2, manual_selection
 
-__all__ = ['kundu_selection_v2', 'manual_selection']
+__all__ = ['kundu_tedpca', 'kundu_selection_v2', 'manual_selection']
